requirements-poc.txt
# Ultra-Minimal Dependencies for POC with Azure OpenAI
# Optimized for cost and performance - removes heavy local ML dependencies

# Core web framework (minimal async)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Azure integrations (replaces local ML models)
azure-identity==1.15.0
openai==1.3.7  # Azure OpenAI support

# Neo4j database
neo4j==5.15.0

# Lightweight text processing (replaces heavy transformers)
langchain-text-splitters==0.0.1  # Only text splitting, no models
tiktoken==0.5.2  # Fast tokenization

# Essential utilities
numpy==1.24.4  # For embeddings array handling
pydantic==2.5.0  # Data validation
python-dotenv==1.0.0  # Environment management

# Minimal logging and monitoring
structlog==23.2.0  # Structured logging
prometheus-client==0.19.0  # Basic metrics

# Security (minimal)
cryptography==41.0.7  # For Azure authentication

# Optional: Development and testing (comment out for production)
# pytest==7.4.3
# pytest-asyncio==0.21.1

# Note: Removed heavy dependencies for cost savings:
# - sentence-transformers (saves ~1GB+ in container size)
# - torch/tensorflow (saves ~2GB+ in container size) 
# - ollama (saves local compute resources)
# - transformers (saves ~500MB+ in container size)
#
# Total container size reduction: ~4GB+ 
# Runtime memory reduction: ~2GB+
# CPU utilization reduction: ~80% (no local embedding inference)